separate kernel routines from class average_pooling_layer (#215)

+++ examples/mnist/test_with_stb_image.cpp
@@ -80,11 +80,11 @@ bool save_image(const std::string& imagefilename,
     )
 {
     // no scaling, save at original size
-    int stride_bytes = img.width();
+    int stride_bytes = (int)img.width();
     int ret = stbi_write_png(
         imagefilename.c_str(),
-        img.width(),
-        img.height(),
+        (int)img.width(),
+        (int)img.height(),
         1,
         &(img.at(0, 0)),
         stride_bytes);
